refactor(vision-mixer): split builder.rs and transitions.rs into sub-modules
Both files had grown past the 1500-line guideline in CLAUDE.md.
Mechanical split with no behaviour change — same public API at the same paths.

transitions.rs (1563) → transitions/:
  mod.rs (239) — type definitions: TransitionType, TransitionError,
    TransitionController struct, PadTarget, ZHandling, PadAction, plus
    FromStr/Display impls.
  plan.rs (411) — pure plan_transition decision logic + rect_contains
    helper + the 11-case plan-matrix tests.
  controller.rs (955) — impl TransitionController (transition, slide/push/
    dip/fade/cut, morph, animate_input, control-source helpers).

builder.rs (1613) → builder/:
  mod.rs (379) — VisionMixerBuilder + BlockBuilder impl, PipelineParams
    struct + caps helpers, setup_overlay_renderer.
  pipeline_gpu.rs (411) — build_gpu_pipeline.
  pipeline_cpu.rs (422) — build_cpu_pipeline.
  pad_layout.rs (280) — build_pad_properties + initial_pad_geom_for_input.
  audio_meter.rs (166) — per-input + PGM level chains + bus message handler.

External call sites unchanged: crate::gst::transitions::{TransitionController,
TransitionType, PadTarget} and vision_mixer::VisionMixerBuilder resolve to
the same items. Internal items inside builder/ are pub(super) so the split
doesn't widen visibility.

Tests: 457 lib + 1 OpenAPI snapshot + 3 pipeline lifecycle — all green.
Clippy clean.
